interface: moved around post items
