+++ pkg/interface/src/views/landscape/components/Home/Post/PostItem/PostContent.js
